+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - New StartMode enum and StartOptions struct for the start function
+- New WarningOptions for the connect function
 
 ## [0.6.4] - 2018-08-02
 ### Added